+++ packages/components-native/src/Form/components/FormMessage/FormMessage.test.tsx
@@ -1,30 +1,41 @@
 import React from "react";
-import { fireEvent, render } from "@testing-library/react-native";
+import {
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from "@testing-library/react-native";
 import { FormMessage } from ".";
 
 describe("FormMessage", () => {
   it("should render null when there are no message to show", () => {
-    const view = render(<FormMessage />);
-    expect(view.toJSON()).toMatchInlineSnapshot(`null`);
+    render(<FormMessage />);
+    expect(screen.toJSON()).toMatchInlineSnapshot(`null`);
   });
 
-  it("should show the message", () => {
-    const { getByText } = render(<FormMessage />);
+  it("should show the message", async () => {
+    render(<FormMessage />);
 
     const description = "🔥";
     FormMessage.show({ description });
-    expect(getByText(description)).toBeDefined();
+    await waitFor(() => {
+      expect(screen.getByText(description)).toBeDefined();
+    });
   });
 
-  it("should close the message", () => {
-    const { getByText, queryByText } = render(<FormMessage />);
+  it("should close the message", async () => {
+    render(<FormMessage />);
 
     const description = "🌚";
     FormMessage.show({ description });
-    expect(getByText(description)).toBeDefined();
+    await waitFor(() => {
+      expect(screen.getByText(description)).toBeDefined();
+    });
 
     FormMessage.close();
-    expect(queryByText(description)).toBeNull();
+    await waitFor(() => {
+      expect(screen.queryByText(description)).toBeNull();
+    });
   });
 
   describe("Opening another message through a message", () => {
@@ -40,33 +51,35 @@ describe("FormMessage", () => {
         },
       });
 
-    it("should show the most recent message", () => {
-      const { getByText, queryByText, getByLabelText } = render(
-        <FormMessage />,
-      );
+    it("should show the most recent message", async () => {
+      render(<FormMessage />);
 
       showMessage();
 
-      expect(getByText(firstMessage)).toBeDefined();
-      expect(queryByText(secondMessage)).toBeNull();
+      await waitFor(() => {
+        expect(screen.getByText(firstMessage)).toBeDefined();
+      });
+      expect(screen.queryByText(secondMessage)).toBeNull();
 
-      fireEvent.press(getByLabelText("Click me"));
+      fireEvent.press(screen.getByLabelText("Click me"));
 
-      expect(getByText(secondMessage)).toBeDefined();
-      expect(queryByText(firstMessage)).toBeNull();
+      expect(screen.getByText(secondMessage)).toBeDefined();
+      expect(screen.queryByText(firstMessage)).toBeNull();
     });
 
-    it("should close the most recent message", () => {
-      const { getByText, queryByText, getByLabelText } = render(
-        <FormMessage />,
-      );
+    it("should close the most recent message", async () => {
+      render(<FormMessage />);
 
       showMessage();
-      fireEvent.press(getByLabelText("Click me"));
+      await waitFor(() => {
+        expect(screen.getByText("Click me")).toBeDefined();
+      });
+      fireEvent.press(screen.getByLabelText("Click me"));
       FormMessage.close();
-
-      expect(getByText(firstMessage)).toBeDefined();
-      expect(queryByText(secondMessage)).toBeNull();
+      await waitFor(() => {
+        expect(screen.getByText(firstMessage)).toBeDefined();
+      });
+      expect(screen.queryByText(secondMessage)).toBeNull();
     });
   });
 });